Создаёт модуляцию
